fix(test): isolate photo chooser DOM coverage
