+++ packages/collab-client/test/composer.test.ts
@@ -1,19 +1,8 @@
 import type { AssistantMessage, ImageContent, SessionEntry } from "@oh-my-pi/pi-wire";
-import { afterAll, afterEach, describe, expect, test } from "bun:test";
-import { createElement } from "react";
-import { Composer, photoPromptAcknowledged } from "../upstream/src/components/shell/Composer";
+import { describe, expect, test } from "bun:test";
+import { photoPromptAcknowledged } from "../upstream/src/components/shell/Composer";
 import { recommendedOptionIndex } from "../upstream/src/components/shell/ask-recommendation";
-import type { ActiveTool, GuestClient, GuestSnapshot } from "../upstream/src/lib/client";
-import {
-  type MiniElement,
-  click,
-  mount,
-  query,
-  queryAll,
-  restoreDomGlobals,
-  textOf,
-  unmountAll,
-} from "./react-dom-harness";
+import type { ActiveTool, GuestSnapshot } from "../upstream/src/lib/client";
 
 const QUESTION = "How should ADR-0036 proceed?";
 const OPTIONS = [
@@ -120,14 +109,6 @@ function snapshot(overrides: Partial<GuestSnapshot> = {}): GuestSnapshot {
   };
 }
 
-afterEach(async () => {
-  await unmountAll();
-});
-
-afterAll(() => {
-  restoreDomGlobals();
-});
-
 describe("collaboration ask recommendations", () => {
   test("finds an explicit recommendation on the matching active option", () => {
     const current = snapshot({
@@ -180,56 +161,3 @@ describe("photo prompt acknowledgements", () => {
   });
 });
 
-describe("photo source chooser", () => {
-  test("routes explicit camera and library choices to separate inputs", async () => {
-    const client = {
-      sendAbort(): void {},
-      sendAgentCmd(): void {},
-      sendPrompt(): void {},
-      sendUiResponse(): void {},
-    } as unknown as GuestClient;
-    const tree = await mount(
-      createElement(Composer, {
-        client,
-        snapshot: snapshot({ uiRequest: null, working: false }),
-        embedded: true,
-      }),
-    );
-    const trigger = query(tree.container, "sh-photo-trigger");
-    const cameraInput = query(tree.container, "sh-photo-input-camera");
-    const libraryInput = query(tree.container, "sh-photo-input-library");
-    if (trigger === null || cameraInput === null || libraryInput === null) {
-      throw new Error("photo chooser controls did not render");
-    }
-    expect(trigger.getAttribute("aria-expanded")).toBe("false");
-    expect(cameraInput.getAttribute("capture")).toBe("environment");
-    expect(libraryInput.getAttribute("capture")).toBeNull();
-
-    let cameraClicks = 0;
-    let libraryClicks = 0;
-    (cameraInput as MiniElement & { click(): void }).click = () => {
-      cameraClicks += 1;
-    };
-    (libraryInput as MiniElement & { click(): void }).click = () => {
-      libraryClicks += 1;
-    };
-
-    await click(trigger);
-    expect(trigger.getAttribute("aria-expanded")).toBe("true");
-    const choices = queryAll(tree.container, "sh-photo-source-option");
-    expect(choices.map(choice => textOf(choice))).toEqual([
-      "Take photoOpen rear camera",
-      "Choose existingPhoto library or files",
-    ]);
-    await click(choices[0] as MiniElement);
-    expect(cameraClicks).toBe(1);
-    expect(libraryClicks).toBe(0);
-    expect(query(tree.container, "sh-photo-source-options")).toBeNull();
-
-    await click(trigger);
-    await click(queryAll(tree.container, "sh-photo-source-option")[1] as MiniElement);
-    expect(cameraClicks).toBe(1);
-    expect(libraryClicks).toBe(1);
-    expect(query(tree.container, "sh-photo-source-options")).toBeNull();
-  });
-});
